+++ CHANGES.rst
@@ -31,7 +31,9 @@ Changes:
 
 Fixes:
 ------
-- No change.
+- Fix `Process` deployment using a `WPS-1/2` URL reference defining a ``GetCapabilities`` request to resolve
+  the corresponding ``DescribeProcess`` request if the `Process` ID can be inferred from other known locations
+  (relates to `#11 <https://github.com/crim-ca/weaver/issues/11>`_).
 
 .. _changes_4.18.0: